Add FavoriteComicsRepository and move repositories to data/repository
