+++ app/src/main/java/com/hoc/comicapp/domain/models/FavoriteComic.kt
@@ -1,0 +1,9 @@
+package com.hoc.comicapp.domain.models
+
+import java.util.*
+
+data class FavoriteComic(
+  val title: String,
+  val thumbnail: String,
+  val createdAt: Date
+)
